AS v3.1
Removed the wait-time for the script to start running. Added a check to see if there has been any activity within 5 min. If there hasn't the script wont check for running games on the list. This means that the program is running nothing and only exists becuase MaLa is running. Once someone presses a button the check resumes.